refactor(kafka-explorer): colocate smart and dumb components by feature
Move page (smart) components into the same directory as their
presentational (dumb) counterparts to reduce directory noise and improve
discoverability: brokers-page → brokers/, topics-page → topics/,
topic-detail-page → topic-detail/.
