+++ gravitee-apim-webui-libs/gravitee-kafka-explorer/src/lib/brokers/brokers-page.component.ts
@@ -15,7 +15,7 @@
  */
 import { Component, inject } from '@angular/core';
 
-import { BrokersComponent } from '../brokers/brokers.component';
+import { BrokersComponent } from './brokers.component';
 import { KafkaExplorerStore } from '../services/kafka-explorer-store.service';
 
 @Component({
