+++ gravitee-apim-webui-libs/gravitee-kafka-explorer/src/lib/topic-detail/topic-detail-page.component.html
