+++ gravitee-apim-webui-libs/gravitee-kafka-explorer/src/public-api.ts
@@ -24,9 +24,9 @@ export * from './lib/services/kafka-explorer.service';
 export * from './lib/services/kafka-explorer-store.service';
 export * from './lib/services/kafka-explorer-config.token';
 export * from './lib/brokers/brokers.component';
-export * from './lib/brokers-page/brokers-page.component';
+export * from './lib/brokers/brokers-page.component';
 export * from './lib/topics/topics.component';
-export * from './lib/topics-page/topics-page.component';
+export * from './lib/topics/topics-page.component';
 export * from './lib/topic-detail/topic-detail.component';
-export * from './lib/topic-detail-page/topic-detail-page.component';
+export * from './lib/topic-detail/topic-detail-page.component';
 export * from './lib/pipes/file-size.pipe';
